requirements.txt
setuptools==65.5.0
pip==22.3.1
PyQt5-sip==12.15.0
PyQt5-Qt5==5.15.15
PyQt5==5.15.11
pyserial==3.5
numpy==2.1.1
pyFirmata==1.1.0
pyparsing==3.1.2
setproctitle==1.3.3
wheel==0.43.0
pyqtgraph==0.12.3